Schedule idle function to update combo-boxes from select_led()/select_gain()
We end up calling select_led() and select_gain() from the analyze and
calibrate threads.  We need to make sure that we do not directly access the
the Gtk widgets from within a thread or undefined behavior might occur, so
use GObject.idle_add() to make sure that the update is called from the
GObject main thread.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>