setup.py: Updated python version and switched to setuptools
Updated the python env version to python3. Without this update, you would need the 'python-is-python3' package to be able to run the script (or to create symlinks to python3).
Also imported setup from setuptools instead of distutils.core, because distutils is deprecated (and will be removed starting from python 3.12). This change also brings easy uninstall ('pip uninstall adi_colorimeter') which was not possible with distutils.

Signed-off-by: bia1708 <bianca-maria.popu@analog.com>